+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgload (development version)
 
+* Shimmed `?` now works even if you've renamed the documentation topic (#220).
+
 # pkgload 1.3.0
 
 * `load_all()` now calls `rlang::check_installed()` to prompt whether